requirements.txt
Flask
pydub
openai
line-bot-sdk
python-dotenv
